refactor(agents): extract integration platform operations
