+++ packages/cli/src/modules/agents/agents.service.ts
@@ -75,9 +75,19 @@ import { AGENT_THREAD_PREFIX } from './builder/builder-tool-names';
 import { Agent } from './entities/agent.entity';
 import { ExecutionRecorder } from './execution-recorder';
 import { ChatIntegrationRegistry } from './integrations/agent-chat-integration';
+import { ChatIntegrationActionExecutor } from './integrations/integration-action-executor';
+import { ChatIntegrationContextQueryExecutor } from './integrations/integration-context-query-executor';
+import { IntegrationMessageContextService } from './integrations/integration-message-context.service';
+import {
+	createIntegrationActionTool,
+	createIntegrationContextTool,
+	getIntegrationToolConnectionDescriptors,
+} from './integrations/integration-tools';
 import { syncAgentIntegrations } from './integrations/integrations-sync';
 import { N8NCheckpointStorage } from './integrations/n8n-checkpoint-storage';
 import { N8nMemory } from './integrations/n8n-memory';
+import { createGetEnvironmentTool } from './tools/environment-tool';
+import { createRichInteractionTool } from './integrations/rich-interaction-tool';
 import { composeJsonConfig, decomposeJsonConfig } from './json-config/agent-config-composition';
 import {
 	buildFromJson,
@@ -796,15 +806,7 @@ export class AgentsService {
 		// can't know on its own (current date, instance timezone, day of week)
 		// via a tool call rather than the system prompt — so values that change
 		// per request don't bust system-prompt prompt caching.
-		try {
-			const { createGetEnvironmentTool } = await import('./tools/environment-tool');
-			agent.tool(createGetEnvironmentTool());
-		} catch (toolError) {
-			this.logger.warn('Failed to inject get_environment tool', {
-				agentId,
-				error: toolError instanceof Error ? toolError.message : String(toolError),
-			});
-		}
+		agent.tool(createGetEnvironmentTool());
 
 		// Inject the rich_interaction tool only for platforms that can actually
 		// render its suspend/resume HITL cards. Two gates:
@@ -817,74 +819,34 @@ export class AgentsService {
 		//   - The integration must declare `supportedComponents`. Platforms
 		//     that omit it (e.g. Linear) have explicitly opted out of
 		//     rich_interaction.
-		const integration = integrationType
-			? Container.get(ChatIntegrationRegistry).get(integrationType)
-			: undefined;
+		const integrationRegistry = Container.get(ChatIntegrationRegistry);
+		const integration = integrationType ? integrationRegistry.get(integrationType) : undefined;
 		if (integration?.supportedComponents !== undefined) {
-			try {
-				const { createRichInteractionTool } = await import('./integrations/rich-interaction-tool');
-				agent.tool(createRichInteractionTool(integrationType));
-			} catch (toolError) {
-				this.logger.warn('Failed to inject rich_interaction tool', {
-					agentId,
-					error: toolError instanceof Error ? toolError.message : String(toolError),
-				});
-			}
+			agent.tool(createRichInteractionTool(integrationType));
 		}
 
 		if (credentialIntegrations.length > 0) {
-			try {
-				const {
-					createIntegrationActionTool,
-					createIntegrationContextTool,
-					getIntegrationToolConnectionDescriptors,
-				} = await import('./integrations/integration-tools');
-				const { IntegrationMessageContextService } = await import(
-					'./integrations/integration-message-context.service'
-				);
-				const { ChatIntegrationActionExecutor } = await import(
-					'./integrations/integration-action-executor'
-				);
-				const { ChatIntegrationContextQueryExecutor } = await import(
-					'./integrations/integration-context-query-executor'
-				);
+			const messageContextStore = Container.get(IntegrationMessageContextService);
+			const actionExecutor = Container.get(ChatIntegrationActionExecutor);
+			const queryExecutor = Container.get(ChatIntegrationContextQueryExecutor);
 
-				const messageContextStore = Container.get(IntegrationMessageContextService);
-				const actionExecutor = Container.get(ChatIntegrationActionExecutor);
-				const queryExecutor = Container.get(ChatIntegrationContextQueryExecutor);
-				const integrationRegistry = Container.get(ChatIntegrationRegistry);
-
-				for (const descriptor of getIntegrationToolConnectionDescriptors(
-					credentialIntegrations,
-					agentId,
-					(integrationConfig) => {
-						const integration = integrationRegistry.get(integrationConfig.type);
-						return {
-							contextQueries: integration?.contextQueries,
-							actions: integration?.actions,
-						};
-					},
-				)) {
-					agent.tool(
-						createIntegrationContextTool({
-							descriptor,
-							messageContextStore,
-							queryExecutor,
-						}),
-					);
-					agent.tool(
-						createIntegrationActionTool({
-							descriptor,
-							messageContextStore,
-							actionExecutor,
-						}),
-					);
-				}
-			} catch (toolError) {
-				this.logger.warn('Failed to inject integration context/action tools', {
-					agentId,
-					error: toolError instanceof Error ? toolError.message : String(toolError),
-				});
+			for (const descriptor of getIntegrationToolConnectionDescriptors(
+				credentialIntegrations,
+				agentId,
+				(integrationConfig) => {
+					const integrationDef = integrationRegistry.get(integrationConfig.type);
+					return {
+						contextQueries: integrationDef?.contextQueries,
+						actions: integrationDef?.actions,
+					};
+				},
+			)) {
+				agent.tool(
+					createIntegrationContextTool({ descriptor, messageContextStore, queryExecutor }),
+				);
+				agent.tool(
+					createIntegrationActionTool({ descriptor, messageContextStore, actionExecutor }),
+				);
 			}
 		}
