+++ packages/cli/src/modules/agents/integrations/integration-context-query-executor.ts
@@ -1,56 +1,26 @@
 import { Service } from '@n8n/di';
-import { z } from 'zod';
 
+import { ChatIntegrationRegistry } from './agent-chat-integration';
 import { ChatIntegrationService } from './chat-integration.service';
-import { normalizePlatformId } from './integration-platform-id';
+import { INTEGRATION_ERROR_CODES } from './integration-error-codes';
+import { connectionUnavailable, integrationError } from './integration-helpers';
 import type {
 	IntegrationContextQuery,
 	IntegrationContextQueryExecutor,
 	IntegrationToolConnectionDescriptor,
 } from './integration-tools';
 
-const getUserInputSchema = z.object({
-	userId: z.string().min(1),
-});
-
-const getChannelInfoInputSchema = z.object({
-	channelId: z.string().min(1),
-});
-
-const searchUsersInputSchema = z
-	.object({
-		query: z.string().min(1).optional(),
-		email: z.string().min(1).optional(),
-		limit: z.number().int().min(1).max(50).default(10),
-		includeBots: z.boolean().default(false),
-		includeDeleted: z.boolean().default(false),
-	})
-	.refine((input) => input.query !== undefined || input.email !== undefined, {
-		message: 'Provide query or email.',
-	});
-
-const searchChannelsInputSchema = z.object({
-	query: z.string().min(1),
-	limit: z.number().int().min(1).max(50).default(10),
-	includeArchived: z.boolean().default(false),
-});
-
-const getIssueInputSchema = z.object({
-	issueId: z.string().min(1),
-	includeComments: z.boolean().default(false),
-	commentsLimit: z.number().int().min(1).max(20).default(10),
-});
-
-const searchIssuesInputSchema = z.object({
-	query: z.string().min(1),
-	limit: z.number().int().min(1).max(50).default(10),
-	teamId: z.string().min(1).optional(),
-	includeArchived: z.boolean().default(false),
-});
-
+/**
+ * Thin dispatcher that resolves the platform integration for a descriptor and
+ * delegates the context query to its {@link AgentChatIntegration.executeContextQuery}
+ * implementation. The actual per-platform logic lives in `platforms/*-operations.ts`.
+ */
 @Service()
 export class ChatIntegrationContextQueryExecutor implements IntegrationContextQueryExecutor {
-	constructor(private readonly chatIntegrationService: ChatIntegrationService) {}
+	constructor(
+		private readonly chatIntegrationService: ChatIntegrationService,
+		private readonly integrationRegistry: ChatIntegrationRegistry,
+	) {}
 
 	async execute(params: {
 		descriptor: IntegrationToolConnectionDescriptor;
@@ -58,767 +28,38 @@ export class ChatIntegrationContextQueryExecutor implements IntegrationContextQu
 		input: Record<string, unknown>;
 		persistence?: { threadId: string; resourceId: string };
 	}): Promise<unknown> {
-		if (!params.descriptor.agentId) {
-			return connectionUnavailable();
-		}
+		if (!params.descriptor.agentId) return connectionUnavailable();
+
 		const chat = this.chatIntegrationService.getChatInstance(params.descriptor.agentId, {
 			type: params.descriptor.integration.type,
 			credentialId: params.descriptor.integration.credentialId,
 		});
-		if (!chat) {
-			return connectionUnavailable();
+		if (!chat) return connectionUnavailable();
+
+		const integration = this.integrationRegistry.get(params.descriptor.integration.type);
+		if (!integration?.executeContextQuery) {
+			return integrationError(
+				INTEGRATION_ERROR_CODES.UNSUPPORTED_QUERY,
+				`The ${params.descriptor.integration.type} integration does not support context queries.`,
+			);
 		}
 
 		try {
-			if (params.descriptor.integration.type === 'linear') {
-				const result = await executeLinearQuery(chat, params.query, params.input);
-				if (result !== undefined) return result;
-			}
-
-			if (params.query === 'get_user') {
-				const input = getUserInputSchema.parse(params.input);
-				const user = await chat.getUser(input.userId);
-				return { ok: true, user };
-			}
-
-			if (params.query === 'search_users') {
-				const input = searchUsersInputSchema.parse(params.input);
-				return await searchSlackUsers(chat, input);
-			}
-
-			if (params.query === 'search_channels') {
-				const input = searchChannelsInputSchema.parse(params.input);
-				return await searchSlackChannels(chat, input);
-			}
-
-			if (params.query !== 'get_channel_info') {
-				return {
-					ok: false,
-					error: {
-						code: 'UNSUPPORTED_QUERY',
-						message: `Unsupported context query: ${params.query}`,
-					},
-				};
-			}
-
-			const input = getChannelInfoInputSchema.parse(params.input);
-			const channel = chat.channel(
-				normalizePlatformId(params.descriptor.integration.type, input.channelId),
-			);
-			const channelInfo = await channel.fetchMetadata();
-			return { ok: true, channel: channelInfo };
+			return await integration.executeContextQuery({
+				chat,
+				descriptor: params.descriptor,
+				query: params.query,
+				input: params.input,
+			});
 		} catch (error) {
-			return {
-				ok: false,
-				error: {
-					code: 'CONTEXT_QUERY_FAILED',
-					message: error instanceof Error ? error.message : String(error),
-				},
-			};
-		}
-	}
-}
-
-function connectionUnavailable() {
-	return {
-		ok: false,
-		error: {
-			code: 'CONNECTION_NOT_AVAILABLE',
-			message: 'The integration connection is not currently available.',
-		},
-	};
-}
-
-type SearchUsersInput = z.infer<typeof searchUsersInputSchema>;
-type SearchChannelsInput = z.infer<typeof searchChannelsInputSchema>;
-type GetIssueInput = z.infer<typeof getIssueInputSchema>;
-type SearchIssuesInput = z.infer<typeof searchIssuesInputSchema>;
-
-interface LinearAdapter {
-	client: Record<string, unknown>;
-}
-
-interface LinearClientWithUser {
-	user(userId: string): Promise<unknown>;
-}
-
-interface LinearClientWithUsers {
-	users(options: Record<string, unknown>): Promise<unknown>;
-}
-
-interface LinearClientWithIssue {
-	issue(issueId: string): Promise<unknown>;
-}
-
-interface LinearClientWithIssueSearch {
-	searchIssues(query: string, options: Record<string, unknown>): Promise<unknown>;
-}
-
-interface SlackUserSearchResult {
-	userId: string;
-	userName: string;
-	fullName: string;
-	email?: string;
-	avatarUrl?: string;
-	isBot: boolean;
-}
-
-interface SlackChannelSearchResult {
-	channelId: string;
-	name: string;
-	isArchived: boolean;
-	isMember: boolean;
-	isPrivate: boolean;
-	memberCount?: number;
-}
-
-interface SlackProfile {
-	display_name?: unknown;
-	email?: unknown;
-	image_192?: unknown;
-	real_name?: unknown;
-}
-
-interface SlackUser {
-	deleted?: unknown;
-	id?: unknown;
-	is_bot?: unknown;
-	name?: unknown;
-	profile?: SlackProfile;
-	real_name?: unknown;
-}
-
-interface SlackChannel {
-	id?: unknown;
-	is_archived?: unknown;
-	is_member?: unknown;
-	is_private?: unknown;
-	name?: unknown;
-	num_members?: unknown;
-}
-
-interface SlackUsersListResponse {
-	members?: SlackUser[];
-	response_metadata?: { next_cursor?: string };
-}
-
-interface SlackUserLookupResponse {
-	user?: SlackUser;
-}
-
-interface SlackConversationsListResponse {
-	channels?: SlackChannel[];
-	response_metadata?: { next_cursor?: string };
-}
-
-interface SlackUsersApi {
-	list?(options: Record<string, unknown>): Promise<SlackUsersListResponse>;
-	lookupByEmail?(options: Record<string, unknown>): Promise<SlackUserLookupResponse>;
-}
-
-interface SlackConversationsApi {
-	list?(options: Record<string, unknown>): Promise<SlackConversationsListResponse>;
-}
-
-interface SlackWebClient {
-	conversations?: SlackConversationsApi;
-	users?: SlackUsersApi;
-}
-
-interface SlackAdapter {
-	client: SlackWebClient;
-	withToken?: (options: Record<string, unknown>) => Promise<Record<string, unknown>>;
-}
-
-async function executeLinearQuery(
-	chat: {
-		getAdapter(name: string): unknown;
-	},
-	query: IntegrationContextQuery,
-	input: Record<string, unknown>,
-): Promise<unknown | undefined> {
-	if (query === 'get_user') {
-		return await getLinearUser(chat, getUserInputSchema.parse(input));
-	}
-	if (query === 'search_users') {
-		return await searchLinearUsers(chat, searchUsersInputSchema.parse(input));
-	}
-	if (query === 'get_issue') {
-		return await getLinearIssue(chat, getIssueInputSchema.parse(input));
-	}
-	if (query === 'search_issues') {
-		return await searchLinearIssues(chat, searchIssuesInputSchema.parse(input));
-	}
-	return undefined;
-}
-
-async function getLinearUser(
-	chat: {
-		getAdapter(name: string): unknown;
-	},
-	input: { userId: string },
-): Promise<unknown> {
-	const adapter = getLinearAdapter(chat);
-	if (!adapter || !hasLinearUserClient(adapter.client)) {
-		return unsupportedLinearQuery('get_user');
-	}
-
-	const user = await adapter.client.user(input.userId);
-	return { ok: true, user: normalizeLinearUser(user) ?? null };
-}
-
-async function searchLinearUsers(
-	chat: {
-		getAdapter(name: string): unknown;
-	},
-	input: SearchUsersInput,
-): Promise<unknown> {
-	const adapter = getLinearAdapter(chat);
-	if (!adapter || !hasLinearUsersClient(adapter.client)) {
-		return unsupportedLinearQuery('search_users');
-	}
-
-	const response = await adapter.client.users({
-		filter: buildLinearUserSearchFilter(input),
-		first: input.limit,
-		includeArchived: input.includeDeleted,
-		includeDisabled: input.includeDeleted,
-	});
-
-	const users = linearNodes(response).map(normalizeLinearUser).filter(isDefined);
-	return { ok: true, users, resultCount: users.length };
-}
-
-async function getLinearIssue(
-	chat: {
-		getAdapter(name: string): unknown;
-	},
-	input: GetIssueInput,
-): Promise<unknown> {
-	const adapter = getLinearAdapter(chat);
-	if (!adapter || !hasLinearIssueClient(adapter.client)) {
-		return unsupportedLinearQuery('get_issue');
-	}
-
-	const issue = await adapter.client.issue(input.issueId);
-	if (!issue) return { ok: true, issue: null };
-
-	return {
-		ok: true,
-		issue: await normalizeLinearIssue(issue, {
-			includeComments: input.includeComments,
-			commentsLimit: input.commentsLimit,
-		}),
-	};
-}
-
-async function searchLinearIssues(
-	chat: {
-		getAdapter(name: string): unknown;
-	},
-	input: SearchIssuesInput,
-): Promise<unknown> {
-	const adapter = getLinearAdapter(chat);
-	if (!adapter || !hasLinearIssueSearchClient(adapter.client)) {
-		return unsupportedLinearQuery('search_issues');
-	}
-
-	const response = await adapter.client.searchIssues(input.query, {
-		first: input.limit,
-		...(input.teamId ? { teamId: input.teamId } : {}),
-		includeArchived: input.includeArchived,
-	});
-
-	const issues = await Promise.all(
-		linearNodes(response).map(async (issue) => await normalizeLinearIssue(issue)),
-	);
-
-	return {
-		ok: true,
-		issues,
-		resultCount: issues.length,
-		...(numberValue(response, 'totalCount') !== undefined
-			? { totalCount: numberValue(response, 'totalCount') }
-			: {}),
-	};
-}
-
-function getLinearAdapter(chat: { getAdapter(name: string): unknown }): LinearAdapter | undefined {
-	const adapter = chat.getAdapter('linear');
-	if (!isRecord(adapter) || !isRecord(adapter.client)) return undefined;
-	return { client: adapter.client };
-}
-
-function hasLinearUserClient(
-	client: Record<string, unknown>,
-): client is Record<string, unknown> & LinearClientWithUser {
-	return typeof client.user === 'function';
-}
-
-function hasLinearUsersClient(
-	client: Record<string, unknown>,
-): client is Record<string, unknown> & LinearClientWithUsers {
-	return typeof client.users === 'function';
-}
-
-function hasLinearIssueClient(
-	client: Record<string, unknown>,
-): client is Record<string, unknown> & LinearClientWithIssue {
-	return typeof client.issue === 'function';
-}
-
-function hasLinearIssueSearchClient(
-	client: Record<string, unknown>,
-): client is Record<string, unknown> & LinearClientWithIssueSearch {
-	return typeof client.searchIssues === 'function';
-}
-
-function unsupportedLinearQuery(
-	query: 'get_user' | 'search_users' | 'get_issue' | 'search_issues',
-) {
-	return {
-		ok: false,
-		error: {
-			code: 'UNSUPPORTED_QUERY',
-			message: `The active Linear connection does not support ${query}.`,
-		},
-	};
-}
-
-function buildLinearUserSearchFilter(input: SearchUsersInput): Record<string, unknown> {
-	const userSearchFilter = buildLinearUserSearchTermFilter(input);
-	if (input.includeBots) return userSearchFilter;
-
-	return {
-		and: [userSearchFilter, { app: { eq: false } }],
-	};
-}
-
-function buildLinearUserSearchTermFilter(input: SearchUsersInput): Record<string, unknown> {
-	if (input.email) {
-		return { email: { eqIgnoreCase: input.email.trim() } };
-	}
-
-	const query = input.query?.trim() ?? '';
-	return {
-		or: [
-			{ name: { containsIgnoreCase: query } },
-			{ displayName: { containsIgnoreCase: query } },
-			{ email: { containsIgnoreCase: query } },
-		],
-	};
-}
-
-export async function normalizeLinearIssue(
-	value: unknown,
-	options: { includeComments?: boolean; commentsLimit?: number } = {},
-): Promise<Record<string, unknown>> {
-	const issue = isRecord(value) ? value : {};
-	const [state, assignee, creator, team, project, labelsConnection] = await Promise.all([
-		awaitableProperty(issue, 'state'),
-		awaitableProperty(issue, 'assignee'),
-		awaitableProperty(issue, 'creator'),
-		awaitableProperty(issue, 'team'),
-		awaitableProperty(issue, 'project'),
-		callLinearConnection(issue, 'labels', { first: 50 }),
-	]);
-
-	const priority = normalizeLinearPriority(issue);
-	const comments = options.includeComments
-		? await normalizeLinearComments(issue, options.commentsLimit ?? 10)
-		: undefined;
-
-	return removeUndefinedValues({
-		issueId: stringProperty(issue, 'id'),
-		identifier: stringProperty(issue, 'identifier'),
-		title: stringProperty(issue, 'title'),
-		description: stringProperty(issue, 'description'),
-		url: stringProperty(issue, 'url'),
-		priority,
-		state: normalizeLinearState(state),
-		assignee: normalizeLinearUser(assignee),
-		creator: normalizeLinearUser(creator),
-		team: normalizeLinearTeam(team),
-		project: normalizeLinearProject(project),
-		labels: linearNodes(labelsConnection).map(normalizeLinearLabel).filter(isDefined),
-		createdAt: isoDateProperty(issue, 'createdAt'),
-		updatedAt: isoDateProperty(issue, 'updatedAt'),
-		...(comments ? { comments } : {}),
-	});
-}
-
-async function normalizeLinearComments(
-	issue: Record<string, unknown>,
-	commentsLimit: number,
-): Promise<Array<Record<string, unknown>> | undefined> {
-	const commentsConnection = await callLinearConnection(issue, 'comments', {
-		first: commentsLimit,
-	});
-	if (!commentsConnection) return undefined;
-	return await Promise.all(
-		linearNodes(commentsConnection).map(async (comment) => await normalizeLinearComment(comment)),
-	);
-}
-
-export async function normalizeLinearComment(value: unknown): Promise<Record<string, unknown>> {
-	const comment = isRecord(value) ? value : {};
-	const author = await awaitableProperty(comment, 'user');
-	return removeUndefinedValues({
-		commentId: stringProperty(comment, 'id'),
-		body: stringProperty(comment, 'body'),
-		url: stringProperty(comment, 'url'),
-		createdAt: isoDateProperty(comment, 'createdAt'),
-		updatedAt: isoDateProperty(comment, 'updatedAt'),
-		author: normalizeLinearUser(author),
-	});
-}
-
-function normalizeLinearUser(value: unknown): Record<string, unknown> | undefined {
-	if (!isRecord(value)) return undefined;
-	const userId = stringProperty(value, 'id');
-	if (!userId) return undefined;
-
-	return removeUndefinedValues({
-		userId,
-		name: stringProperty(value, 'name') ?? userId,
-		displayName: stringProperty(value, 'displayName'),
-		email: stringProperty(value, 'email'),
-		avatarUrl: stringProperty(value, 'avatarUrl'),
-		active: booleanProperty(value, 'active'),
-		isBot: booleanProperty(value, 'app') ?? false,
-		isAssignable: booleanProperty(value, 'isAssignable'),
-		isMentionable: booleanProperty(value, 'isMentionable'),
-		url: stringProperty(value, 'url'),
-	});
-}
-
-function normalizeLinearState(value: unknown): Record<string, unknown> | undefined {
-	if (!isRecord(value)) return undefined;
-	return removeUndefinedValues({
-		id: stringProperty(value, 'id'),
-		name: stringProperty(value, 'name'),
-		type: stringProperty(value, 'type'),
-	});
-}
-
-function normalizeLinearTeam(value: unknown): Record<string, unknown> | undefined {
-	if (!isRecord(value)) return undefined;
-	return removeUndefinedValues({
-		teamId: stringProperty(value, 'id'),
-		key: stringProperty(value, 'key'),
-		name: stringProperty(value, 'name'),
-	});
-}
-
-function normalizeLinearProject(value: unknown): Record<string, unknown> | undefined {
-	if (!isRecord(value)) return undefined;
-	return removeUndefinedValues({
-		projectId: stringProperty(value, 'id'),
-		name: stringProperty(value, 'name'),
-		url: stringProperty(value, 'url'),
-	});
-}
-
-function normalizeLinearLabel(value: unknown): Record<string, unknown> | undefined {
-	if (!isRecord(value)) return undefined;
-	const labelId = stringProperty(value, 'id');
-	const name = stringProperty(value, 'name');
-	if (!labelId || !name) return undefined;
-	return { labelId, name };
-}
-
-function normalizeLinearPriority(
-	issue: Record<string, unknown>,
-): Record<string, unknown> | undefined {
-	const value = numberProperty(issue, 'priority');
-	const label = stringProperty(issue, 'priorityLabel');
-	if (value === undefined && !label) return undefined;
-	return removeUndefinedValues({ value, label });
-}
-
-async function awaitableProperty(
-	record: Record<string, unknown>,
-	key: string,
-): Promise<unknown | undefined> {
-	return await Promise.resolve(record[key]);
-}
-
-async function callLinearConnection(
-	record: Record<string, unknown>,
-	key: string,
-	options: Record<string, unknown>,
-): Promise<unknown | undefined> {
-	const fn = record[key];
-	if (!isLinearConnectionFunction(fn)) return undefined;
-	return await fn.call(record, options);
-}
-
-function linearNodes(value: unknown): unknown[] {
-	if (!isRecord(value) || !Array.isArray(value.nodes)) return [];
-	return value.nodes;
-}
-
-function removeUndefinedValues(value: Record<string, unknown>): Record<string, unknown> {
-	return Object.fromEntries(Object.entries(value).filter(([, entry]) => entry !== undefined));
-}
-
-function isDefined<T>(value: T | undefined): value is T {
-	return value !== undefined;
-}
-
-function stringProperty(record: Record<string, unknown>, key: string): string | undefined {
-	return stringValue(record[key]);
-}
-
-function numberProperty(record: Record<string, unknown>, key: string): number | undefined {
-	const value = record[key];
-	return typeof value === 'number' ? value : undefined;
-}
-
-function numberValue(value: unknown, key: string): number | undefined {
-	return isRecord(value) ? numberProperty(value, key) : undefined;
-}
-
-function booleanProperty(record: Record<string, unknown>, key: string): boolean | undefined {
-	const value = record[key];
-	return typeof value === 'boolean' ? value : undefined;
-}
-
-function isoDateProperty(record: Record<string, unknown>, key: string): string | undefined {
-	const value = record[key];
-	if (value instanceof Date) return value.toISOString();
-	if (typeof value !== 'string' || value.length === 0) return undefined;
-	const date = new Date(value);
-	return Number.isNaN(date.getTime()) ? value : date.toISOString();
-}
-
-type LinearConnectionFunction = (
-	this: Record<string, unknown>,
-	options: Record<string, unknown>,
-) => unknown | PromiseLike<unknown>;
-
-function isLinearConnectionFunction(value: unknown): value is LinearConnectionFunction {
-	return typeof value === 'function';
-}
-
-async function searchSlackUsers(
-	chat: {
-		getAdapter(name: string): unknown;
-	},
-	input: SearchUsersInput,
-): Promise<unknown> {
-	const adapter = getSlackAdapter(chat);
-	const usersApi = adapter?.client.users;
-	if (!adapter || !usersApi?.list) {
-		return unsupportedSlackSearch('search_users');
-	}
-
-	const usersById = new Map<string, SlackUserSearchResult>();
-	const searchTerm = normalizeSearchTerm(input.query ?? input.email ?? '');
-	const email = input.email?.trim();
-
-	if (email && adapter.client.users?.lookupByEmail) {
-		try {
-			const response = await adapter.client.users.lookupByEmail(
-				await withSlackToken(adapter, { email }),
+			return integrationError(
+				INTEGRATION_ERROR_CODES.CONTEXT_QUERY_FAILED,
+				error instanceof Error ? error.message : String(error),
 			);
-			const user = normalizeSlackUser(response.user);
-			if (
-				user &&
-				shouldIncludeUser(response.user, input) &&
-				matchesUserSearch(response.user, searchTerm, email)
-			) {
-				usersById.set(user.userId, user);
-			}
-		} catch {
-			// Fall back to users.list, which can still match workspaces/scopes that
-			// do not support users.lookupByEmail for this token.
 		}
 	}
-
-	let cursor: string | undefined;
-	do {
-		const response = await usersApi.list(
-			await withSlackToken(adapter, { limit: 200, ...(cursor ? { cursor } : {}) }),
-		);
-		for (const rawUser of response.members ?? []) {
-			if (!shouldIncludeUser(rawUser, input) || !matchesUserSearch(rawUser, searchTerm, email)) {
-				continue;
-			}
-			const user = normalizeSlackUser(rawUser);
-			if (user) usersById.set(user.userId, user);
-			if (usersById.size >= input.limit) break;
-		}
-		cursor = response.response_metadata?.next_cursor || undefined;
-	} while (cursor && usersById.size < input.limit);
-
-	const users = [...usersById.values()].slice(0, input.limit);
-	return { ok: true, users, resultCount: users.length };
 }
 
-async function searchSlackChannels(
-	chat: {
-		getAdapter(name: string): unknown;
-	},
-	input: SearchChannelsInput,
-): Promise<unknown> {
-	const adapter = getSlackAdapter(chat);
-	const conversationsApi = adapter?.client.conversations;
-	if (!adapter || !conversationsApi?.list) {
-		return unsupportedSlackSearch('search_channels');
-	}
-
-	const channelsById = new Map<string, SlackChannelSearchResult>();
-	const searchTerm = normalizeChannelSearchTerm(input.query);
-	let cursor: string | undefined;
-
-	do {
-		const response = await conversationsApi.list(
-			await withSlackToken(adapter, {
-				exclude_archived: !input.includeArchived,
-				limit: 200,
-				types: 'public_channel,private_channel',
-				...(cursor ? { cursor } : {}),
-			}),
-		);
-
-		for (const rawChannel of response.channels ?? []) {
-			if (!input.includeArchived && rawChannel.is_archived === true) continue;
-			if (!matchesChannelSearch(rawChannel, searchTerm)) continue;
-			const channel = normalizeSlackChannel(rawChannel);
-			if (channel) channelsById.set(channel.channelId, channel);
-			if (channelsById.size >= input.limit) break;
-		}
-		cursor = response.response_metadata?.next_cursor || undefined;
-	} while (cursor && channelsById.size < input.limit);
-
-	const channels = [...channelsById.values()].slice(0, input.limit);
-	return { ok: true, channels, resultCount: channels.length };
-}
-
-function getSlackAdapter(chat: { getAdapter(name: string): unknown }): SlackAdapter | undefined {
-	const adapter = chat.getAdapter('slack');
-	return isSlackAdapter(adapter) ? adapter : undefined;
-}
-
-async function withSlackToken(
-	adapter: SlackAdapter,
-	options: Record<string, unknown>,
-): Promise<Record<string, unknown>> {
-	if (!adapter.withToken) return options;
-	return await adapter.withToken(options);
-}
-
-function unsupportedSlackSearch(query: 'search_users' | 'search_channels') {
-	return {
-		ok: false,
-		error: {
-			code: 'UNSUPPORTED_QUERY',
-			message: `The active Slack connection does not support ${query}.`,
-		},
-	};
-}
-
-function normalizeSlackUser(user: SlackUser | undefined): SlackUserSearchResult | undefined {
-	const userId = stringValue(user?.id);
-	if (!userId) return undefined;
-	const profile = user?.profile;
-	const userName = stringValue(user?.name) ?? userId;
-	const fullName =
-		stringValue(user?.real_name) ??
-		stringValue(profile?.real_name) ??
-		stringValue(profile?.display_name) ??
-		userName;
-
-	return {
-		userId,
-		userName,
-		fullName,
-		...(stringValue(profile?.email) ? { email: stringValue(profile?.email) } : {}),
-		...(stringValue(profile?.image_192) ? { avatarUrl: stringValue(profile?.image_192) } : {}),
-		isBot: user?.is_bot === true,
-	};
-}
-
-function normalizeSlackChannel(
-	channel: SlackChannel | undefined,
-): SlackChannelSearchResult | undefined {
-	if (!channel) return undefined;
-	const channelId = stringValue(channel?.id);
-	const name = stringValue(channel?.name);
-	if (!channelId || !name) return undefined;
-
-	return {
-		channelId,
-		name: `#${name}`,
-		isArchived: channel.is_archived === true,
-		isMember: channel.is_member === true,
-		isPrivate: channel.is_private === true,
-		...(typeof channel.num_members === 'number' ? { memberCount: channel.num_members } : {}),
-	};
-}
-
-function shouldIncludeUser(user: SlackUser | undefined, input: SearchUsersInput): boolean {
-	if (!input.includeDeleted && user?.deleted === true) return false;
-	if (!input.includeBots && user?.is_bot === true) return false;
-	return true;
-}
-
-function matchesUserSearch(
-	user: SlackUser | undefined,
-	searchTerm: string,
-	email: string | undefined,
-): boolean {
-	const fields = [
-		stringValue(user?.id),
-		stringValue(user?.name),
-		stringValue(user?.real_name),
-		stringValue(user?.profile?.display_name),
-		stringValue(user?.profile?.real_name),
-		stringValue(user?.profile?.email),
-	]
-		.filter((field): field is string => field !== undefined)
-		.map(normalizeSearchTerm);
-
-	if (email) {
-		const normalizedEmail = normalizeSearchTerm(email);
-		if (fields.some((field) => field === normalizedEmail)) return true;
-	}
-
-	return fields.some((field) => field.includes(searchTerm));
-}
-
-function matchesChannelSearch(channel: SlackChannel | undefined, searchTerm: string): boolean {
-	const fields = [stringValue(channel?.id), stringValue(channel?.name)]
-		.filter((field): field is string => field !== undefined)
-		.map(normalizeChannelSearchTerm);
-	return fields.some((field) => field.includes(searchTerm));
-}
-
-function normalizeSearchTerm(value: string): string {
-	return value.trim().toLowerCase().replace(/^@/, '');
-}
-
-function normalizeChannelSearchTerm(value: string): string {
-	return value.trim().toLowerCase().replace(/^#/, '');
-}
-
-function stringValue(value: unknown): string | undefined {
-	return typeof value === 'string' && value.length > 0 ? value : undefined;
-}
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-	return typeof value === 'object' && value !== null;
-}
-
-function isSlackAdapter(value: unknown): value is SlackAdapter {
-	if (!isRecord(value) || !isRecord(value.client)) return false;
-	if (
-		'withToken' in value &&
-		value.withToken !== undefined &&
-		typeof value.withToken !== 'function'
-	) {
-		return false;
-	}
-	return true;
-}
+// Re-export the Linear normalizers used by tests / action executor message-context
+// construction. Tests reference these via this module path.
+export { normalizeLinearComment, normalizeLinearIssue } from './platforms/linear-operations';
