--- packages/cli/src/modules/agents/integrations/integration-platform-id.ts
@@ -1,3 +1,0 @@
-export function normalizePlatformId(platform: string, id: string): string {
-	return id.includes(':') ? id : `${platform}:${id}`;
-}
